pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "remsfal-ocr"
version = "0.1.0"
description = "REMSFAL OCR Microservice - A stateless Python service for text extraction"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "REMSFAL Team"}
]
keywords = ["ocr", "text-extraction", "kafka", "minio", "paddleocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "azure-storage-blob>=12.27.1",
    "azure-identity>=1.25.1",
    "azure-keyvault-secrets>=4.10.0",
    "kafka-python==2.3.0",
    "minio==7.2.20",
    "paddleocr==2.10.0",
    "paddlepaddle==3.2.2",
    "pydantic==2.10.6",
    "python-dotenv==1.2.1",
    "setuptools==80.9.0",
]

[project.optional-dependencies]
test = [
    "pytest==8.4.2",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "testcontainers[kafka,minio]>=4.0.0",
]
lint = [
    "flake8>=7.0.0",
]
dev = [
    "pytest==8.4.2",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "testcontainers[kafka,minio]>=4.0.0",
    "flake8>=7.0.0",
]

[tool.hatch.envs.default]
features = ["dev"]

[tool.hatch.envs.default.scripts]
test = "pytest test/ {args}"
test-cov = "pytest test/ --cov=src --cov-report=term-missing {args}"
test-cov-full = "pytest test/ --cov=src --cov-report=lcov --cov-report=term-missing --cov-report=html --cov-report=xml {args}"
lint = "flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics"
lint-report = "flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics --output-file=report/flake8-report.txt"
check = [
    "lint",
    "test-cov",
]
start = "python src/main.py"

[tool.hatch.envs.test]
features = ["test"]

[[tool.hatch.envs.test.matrix]]
python = ["3.11", "3.12", "3.13"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/test/*",
    "*/tests/*", 
    "*/__pycache__/*",
    "*/venv/*",
    "*/virtualenv/*",
    "*/.tox/*",
    "*/site-packages/*"
]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod"
]

[tool.coverage.html]
directory = "coverage/html"

[tool.coverage.xml]
output = "coverage/coverage.xml"

[tool.coverage.lcov]
output = "coverage/coverage.lcov"